+++ indexer/services/roundtable/__tests__/tasks/create-pnl-ticks.test.ts
@@ -11,7 +11,7 @@ import {
   FundingIndexUpdatesTable,
 } from '@dydxprotocol-indexer/postgres';
 
-import createPnlTicksTask, { normalizeStartTime } from '../../src/tasks/create-pnl-ticks';
+import createPnlTicksTask from '../../src/tasks/create-pnl-ticks';
 import { LatestAccountPnlTicksCache, PnlTickForSubaccounts, redis } from '@dydxprotocol-indexer/redis';
 import { DateTime } from 'luxon';
 import config from '../../src/config';
@@ -121,18 +121,6 @@ describe('create-pnl-ticks', () => {
         },
       ]),
     );
-  });
-
-  it('normalizeStartTime', () => {
-    const time: Date = new Date('2021-01-09T20:00:50.000Z');
-    // 1 hour
-    config.PNL_TICK_UPDATE_INTERVAL_MS = 1000 * 60 * 60;
-    const result1: Date = normalizeStartTime(time);
-    expect(result1.toISOString()).toBe('2021-01-09T20:00:00.000Z');
-    // 1 day
-    config.PNL_TICK_UPDATE_INTERVAL_MS = 1000 * 60 * 60 * 24;
-    const result2: Date = normalizeStartTime(time);
-    expect(result2.toISOString()).toBe('2021-01-09T00:00:00.000Z');
   });
 
   it('succeeds with no prior pnl ticks and open perpetual positions', async () => {
